api/requirements.txt
tensorflow>=2.20.0
fastapi>=0.104.1
uvicorn>=0.24.0
python-multipart>=0.0.6
Pillow>=10.1.0
numpy>=1.24.3
python-json-logger>=2.0.7
requests>=2.31.0